software/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brswdiy-app"
version = "0.1.0"
description = "Desktop app for BRSWDIY"
requires-python = ">=3.11"
dependencies = [
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "black>=24.0",
    "ruff>=0.6.0",
]
